checkup: Create components
This change introduce the logic for creating each of checkup
components (e.g: Namespace, Job, etc..) in the cluster.

The creation of the components is performed in the Setup method.
In case of failure, it will attempt to cleanup any intermediate objects.
The clean up is done by deleting the Namespace and eventually all the objects
inside it [1].

The kubernetes fake client used in tests records objects that are created and
deleted.
In order to assert the results correctly, a normalization is required to some of
the empty containers (slices, maps) the fake tracks.
Any 'nil' value is converted to an initialized empty container, allowing a semantic
equality to pass.

[1] https://kubernetes.io/docs/concepts/architecture/garbage-collection/#background-deletion

Signed-off-by: Or Mergi <ormergi@redhat.com>