checkup: Delete components at teardown
This commit introduce the logic for removing the checkup related objects
from the cluster that was created by checkup.Setup().

checkup.Teardown() deletes the checkup related objects by deleting
the Namespace, that eventually triggers deletion of all the objects inside
it [1].
ClusterRoleBindings are deleted explicitly as they are not namespaced
objects.

[1] https://kubernetes.io/docs/concepts/architecture/garbage-collection/#background-deletion

Signed-off-by: Or Mergi <ormergi@redhat.com>