reporter: Add reporting logic
This adds the reporter main logic - the Report() method.
The method reports the reportee's state to the user-defined ConfigMap
object, which also contains the checkup's configuration / spec.

This also introduces the Reportee interface which enables the reporter
to query the reportee object (the checkup) for its current state,
so it could be reported to the user.

The required empty methods are added to the Checkup struct,
in order to make it implement the Reportee interface.

Signed-off-by: Orel Misan <omisan@redhat.com>