automation: Use defaults when no args are provided
The current default handling in the automation scripts are not scalable.
Each time a new argument is added, there is a need to also add it
to the default handling condition.

With this change, instead of checking if each specific argument is not
set, a global check is done against the number of arguments the script
is called with. If no arguments are provided, the default options are
triggered.

Signed-off-by: Edward Haas <edwardh@redhat.com>