results: Read and report checkup mandatory results
Until now, the results reported by the checkup hasn't been read
by the framework.

This commit adds reading the checkup results, and consider them
in the final report to the user.

The "succeeded" field value is a "true", when no errors had occurred
in the framework or in the checkup.

The "failureReason" field is a concatenation of:
1. The value reported by the checkup (if exists).
2. Checkup.Run() error (if occurred).
3. Checkup.Results() error (if occurred).
4. Checkup.Teardown() error (if occurred).

checkupStub's methods now have a pointer receiver, as for the linter's
suggestion, due to its size.

Signed-off-by: Orel Misan <omisan@redhat.com>