launcher: Report succeeded and failureReason
Replace the hard codded report of "succeeded" and "failureReason"
with an actual report.

The "succeeded" field is reported as "true" only when:
Checkup.Setup(), Checkup.Run() and Checkup.Teardown()
are successfully completed.

At the moment, because the checkup results are not read,
beyond the returned error from Checkup.Run(), there is not enough
data to declare the checkup as failed.

Change the Launcher's tests, to test the actual final report
of these fields (success scenario).

Signed-off-by: Orel Misan <omisan@redhat.com>