crds: Add golang code types and gen infra
The CRDs can be generated with controller-gen. This change add a new
automation option `--gen-crds` that will read api/ and generate them at
deploy/crds/.

To handle the API as a CRD they need to have deep copy fuctions so they
can be used by k8s golang libraries like the other resources. This
change add the controller-gen to generate them.

To make the checkups input parameters more flexible this change convert
the EnvVars input from the CR into json.RawMessage data field that can be
populated as a yaml/json object, the map entries will be passed as env
vars to the checkup.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>