launcher: Move the error collection out of the defer
Collecting the errors in the defer closures requires a condition which
can be avoided.

Collecting the errors at the function body, just before returning the
error, simplifies things.
The disadvantage is the need to do so on each `return` with an error.
As this function steps are not expected to change much, this seems a
reasonable compromise when comparing with the logic residing in the
defer.

Signed-off-by: Edward Haas <edwardh@redhat.com>