reporter: Do not use pointers when not needed
When function receivers or arguments are defined as pointers, beyond the
functional need it also exposes an intent.

When reading the code, an expression like `foo(&myvar)` has the
potential to be interpreted as: "foo is mutating myvar".
On the other hand, seeing `foo(myvar)` is more likely to express to the
reader that the intent is to keep the variable immutable.

The change reverts 6b9e203eac3fd092cfac90a106dbfab20e4099b8

Signed-off-by: Edward Haas <edwardh@redhat.com>