checkup: Replace client type to kubernetes.Interface
Until now, Checkup held as a state, a custom "client" interface
which was a small subset of the default "kubernetes.Interface"
client set.

The custom client interface was set up because we didn't
know in advanced if K8s fake client will be suitable for our needs.

The custom client interface was forcing its users to use
a specific client version (for example CoreV1) instead of having
a choice to select a client from the default client set.

This had a system-wide affect, affecting all functions that were called
from Checkup, and using the client.

Because the "configmap" package was affected, other packages were
affected as well.

Holding a client from type "kubernetes.Interface", allows its users
a choice to select the most suitable client.

Signed-off-by: Orel Misan <omisan@redhat.com>