kubevirt-vm-latency: Delete VMs on teardown
Teardown() does best effort to delete the objects created by Setup() (VMIs).
It deletes the VMIs and wait for them to dispose.
In case a VM failed to delete or exceeded disposal timeout, it will try the next one.
Finally, it return an error if one or more VMIs failed delete/ dispose.

It tests a context with short timeout is passed to Teardown() in order to
override the default timeout.

Notes about 'launcher' tests,
- Since Kubevirt client-go wont provide a fake-client for VMIs, it was necessary
  to implement one with most basic functionallity that can supports tracking VMIs.

- On VMI create it will append the VMI Ready condition, in order to prevent from
  'Launcher{}' to hang while waiting for the VMIs to become ready.

Signed-off-by: Or Mergi <ormergi@redhat.com>