config: Add check if ConfigMap is already in use
At the moment, it is possible to run two (or more) concurrent
framework Jobs, which refer to the same ConfigMap object as an input.
The results of such a scenario are undefined because there will be
a race when writing the results.

It is also possible to run two (or more) consecutive framework Jobs,
which refer to the same ConfigMap object as an input.
In this case, the results of the first run will be overwritten with the
results of the second run.

Add a check if the input ConfigMap object is already in use by another
checkup Job, and fail if it is.
We check whether the "StartTimestamp" field reported by the reporter
exists in the input.

Signed-off-by: Orel Misan <omisan@redhat.com>