kubevirt-vm-latency: Do not set node-selector when node name is empty
In order to ensure a VM is scheduled on specific node, the VM node
selector is set with "kubernetes.io/hostname" label.

Setting this label with an empty string (i.e: kubernetes.io/hostname="")
cause the VMs to stuck on 'Scheduling' phase as it looks for a node
that does not exists.
It seems that the value of "kubernetes.io/hostname" label is not validated.

Signed-off-by: Or Mergi <ormergi@redhat.com>