checkup, vmlatency: Login to source VMI
The first step in examining the latency (and connection) between two
VMI/s is to access the guest console.

Accessing the console requires:
- Using the kubevirt subresource to reach the console. This is exposed
  as a kubevirt client command.
- Operating in a terminal (e.g. parsing, waiting for expected response).
  This is achieved by using the `goexpect` library.
- Login logic to the respected OS (e.g. Fedora at the moment).
- Triggering the login at the right place and moment.

To support the above, a `console` package now handles all console
operations and a `latency` package that handles the logic of using the
console (which currently just calls login to the source VMI).

The console implementation is mostly taken from the kubevirt [1]
project.

[1] https://github.com/kubevirt/kubevirt/tree/v0.53.0/tests/console

Signed-off-by: Edward Haas <edwardh@redhat.com>