checkups, vmlatency: Support a limited bridge binding
The existing implementation supports only SR-IOV VMI interfaces (and
therefore NetworkAttachementDefinitions that define SR-IOV networks).

This is problematic for testing purposes, therefore, a small hack is
introduced: When the NAD name is `bridge-network`, the VMI is created
with a bridge binding interface.

The hack should be removed when auto detecting the SR-IOV based NAD/s is
introduced.

Signed-off-by: Edward Haas <edwardh@redhat.com>