latency, CI, e2e: Run 'Deploy kiagnose' from main e2e.sh
The --deploy-kiagnose target was performed from the latency
checkup's e2e.sh script.
This target is general to all checkups, so it was moved to
the main e2e.sh script.

Add a build step to kiagnose e2e workflow,
in order to deploy a version of the kiagnose core built from
current code.

Signed-off-by: Orel Misan <omisan@redhat.com>