kiagnose, namespace: Use watcher to wait for deletion
Waiting for the namespace to be deleted was done using polling.
This has now been improved by using a watcher.

The manifest was updated to allow the `watch` operation on namespaces.

In order to test the checkup flows with the new `Until` tooling in
place, it is now necessary to make sure the objects (e.g. namespace) has
the resource-version set, otherwise an error is returned.

Signed-off-by: Edward Haas <edwardh@redhat.com>