kiagnose, checkup: Use k8s client-go `Until` tool for watching the job
Kiagnose creates a checkup job and waits for it to finish.
The current implementation uses a watcher with a self curated logic to
process the events received from it.

The client-go tooling module has a `watch` package [1] that exposes the
function `Until`, implementing all needed mechanics that are required to
process watchers events.

By using this tool, it is now possible to drop our own limited
implementation.
It is worth noting that `Until` also handles possible connectivity
issues between the client and the server, increasing stability in case
of connectivity flakiness.

The change is following the work done for the namespace deletion [2] but
includes an improvements for early termination:
- In case a DELETED event type is detected, fail the waiting earlier.

[1] https://pkg.go.dev/k8s.io/client-go/tools/watch
[2] https://github.com/kiagnose/kiagnose/pull/84

Signed-off-by: Edward Haas <edwardh@redhat.com>