gotestsuite: Introduce a test suite structure for tests
The test suite is providing a structure to write tests.
One is able to add fixtures (currently for setup only) and ref
them from the test itself in case it is needed.

Some fixtures may be added and instructed to run before each test,
without the need to ref them explicitly from the tests.

The checkup teardown tests are adjusted to use the new suite.

Signed-off-by: Edward Haas <edwardh@redhat.com>