vmlatency: Support bridge network
Currently the checkup supports SR-IOV networks only, as it always
define the VMs interface with SR-IOV binding method regardless of the
NetworkAttachmentDefinition (NAD) Spec.Config.

In order to support bridge networks as well and define the VMs correctly,
its necessary to realize which CNI plugin is being used by the NAD and
define the VM interface with the right binding method.

SR-IOV binding method is used when SR-IOV CNI specified in the NAD config.
For other CNIs (e.g: bridge) bridge binding method is used.

Signed-off-by: Or Mergi <ormergi@redhat.com>