checkup: Introduce target Namespace
Currently, checkups are executed in ephemeral Namespace objects.

This change enables checkup execution in a target Namespace -
the Namespace in which the user-supplied ConfigMap object is at.

Create and delete an ephemeral namespace only if the user-supplied
ConfigMap object is at the "kiagnose" namespace.
Else, setup all objects in the target Namespace.

Signed-off-by: Orel Misan <omisan@redhat.com>