checkup: Remove SA on teardown with target namespace
Currently, the teardown removes the whole namespace in which the
objects are living at.

When using a target namespace, there is a need to remove the objects
created by Kiagnose one by one.

Remove the ServiceAccount object created for the checkup by Kiagnose.
Add a permission to delete ServiceAccount objects.

Signed-off-by: Orel Misan <omisan@redhat.com>