checkup: Remove checkup Job on teardown with target namespace
Currently, the teardown removes the whole namespace in which the
objects are living at.

When using a target namespace, there is a need to remove the objects
created by Kiagnose one by one.

Remove the checkup Job object created by Kiagnose.

Use `DeletePropagationForeground` delete policy for the Job in order
to make sure the underlying Pod is removed before going forward with
removing its permissions in the following commits.

Signed-off-by: Orel Misan <omisan@redhat.com>