vm latency checkup, e2e: Run checkup in a target namespace
Currently, the e2e test of VM latency checkup is being executed in an
ephemeral namespace.
In order to show the target namespace feature works, execute the checkup
in a target namespace.

Dropped the test of using the VM latency checkup with ephemeral namespace.
This was done in order to transition this checkup to be used only on
target Namespaces, so its permissions could be lowered in a future PR.

Signed-off-by: Orel Misan <omisan@redhat.com>