checkup, tests: Add setup failure in target ns
Currently, setup failures in a target namespace mode are not
tested.
Add a test of multiple setup failures in target ns mode.

Not checking the state of the namespace after a failure, because
no cleanup logic is in place.

Not checking failure to create ClusterRoleBinding objects because
this logic will be removed in a follow-up PR.

Signed-off-by: Orel Misan <omisan@redhat.com>