checkup, tests: Add teardown failures on target ns
Currently, teardown failures in a target namespace mode are not
tested.

Add a test of multiple teardown failures in target ns mode.

Not checking failure to delete ClusterRoleBindings, because
this logic is about to be removed in a follow-up PR.

Rename `newNamespace()` to `newTestNamespace()` so it won't receive
a namespace name.

Signed-off-by: Orel Misan <omisan@redhat.com>