vmltancy, checkup, tests client: Record created VMs
In tests there is no actual backend behind the stub client
that calculates the VMI status (e.g: with conditions).
Thus its necessary to inject a predefined VMI object so
that test could proceed when trying to fetch a VMI.

With these changes the stub client records created VMIs object
and add the "Ready" condition, and it wont be necessary to inject
predefined VMI object.

Signed-off-by: Or Mergi <ormergi@redhat.com>