vmlatency, checkup: Ensure VMs get scheduled on diffrent nodes
Currently, the source and target node may end up scheduled on
the same node.
Though this may be useful for some use cases it's not the desired
default behavior.

With these changes the checkup setup is changed as follows:
1. Create source VM.
2. Wait for the source VM to be ready.
3. Look into the source VM status and create the target VM
   with the corresponding node anti-affinity rule.
4. Wait for the target VM to be ready.

The source and target will not be scheduled on the same node.

In order to force the VMs to be scheduled on the same node, the
user should specify both source and target node names in the
checkup config.

Signed-off-by: Or Mergi <ormergi@redhat.com>