vmlatency, checkup: Ensure VMs get scheduled on different nodes
Currently, when both the source and the target nodes are not defined
by the user, the source and target VMs may end up scheduled on the
same node.
Though this may be useful for some use cases it's not the desired
default behavior.

The VMs will be created with "latency-check-vm" label which indicates
they belong to the latency checkup.

As long as both "source_node" and "target_node" are not specified in
the checkup config, both VMs will be created with a pod-anti-affinity
[1] rule.
It will ensure that both VMs won't end up running on the same node,
using the "latency-check-vm" label as the labelSelector.

In order to override this behavior and force both VMs to run on the
same node, specify both "source_node" and "target_node" in the checkup
config.
Each VM will be created with correspond node-affinity rule [2].

[1] https://kubernetes.io/docs/concepts/scheduling-eviction/assign-pod-node/#types-of-inter-pod-affinity-and-anti-affinity
[2] https://kubernetes.io/docs/concepts/scheduling-eviction/assign-pod-node/#node-affinity

Signed-off-by: Or Mergi <ormergi@redhat.com>