vmlatency, checkup setup: Wait for VMs IP addresses
Currently, at the checkup setup phase, a VM is considered ready when the
"Ready" condition is present in the VM status.
Run() waits for the VMs IP addresses to present in the VM status, which
indicated that the VM has booted (as the guest-agent reported the IP addresses).

Instead of waiting for different states of the VMs at different times, wait for
the VM IP address to report as part of Setup().
Run() will only console to the VM and start the latency measurement w/o waiting
for any state of the VM.

Signed-off-by: Or Mergi <ormergi@redhat.com>