vmlatency, config: Validate source and target node names
Currently, it's possible to specify one, both or neither of
'source_node' and 'target_node' parameters.
In order to simplify the user experience, the code and tests
the 'source_node' and 'target_node' parameters will be respected
only when both are specified.

Specifying only one of "source_node" and "target_node" is no
longer supported.

Signed-off-by: Or Mergi <ormergi@redhat.com>