vmlatency, checkup: Randomize VMs affinity label
Running more than one instance of kubevirt latency checkup
is not possible due to the fact that VMs of the first instance
will cause the VMs of the other instances to be stuck on the
scheduling phase due to the first instance VMs affinity rules.

Change the VMs label that is used by their Pod anti-affinity to
have a random value, so that VMs that were created by one instance
of the checkup won't block other instances VMs.

Signed-off-by: Or Mergi <ormergi@redhat.com>