checkup: Add suffix to sub-job
In order to enable a namespace admin to execute checkups while
keeping the kiagnose API, there is a need to deploy kiagnose to
the target namespace.

Currently, the Job object created for the checkup by Kiagnose
is named after the user-supplied ConfigMap object's name.
This could cause a naming collision in case the kiagnose Job name is
equal to the user-supplied ConfigMap name.

Add the "-checkup" suffix to the checkup sub-job.

Signed-off-by: Orel Misan <omisan@redhat.com>