kiagnose: Introduce a public environment package
Start using kiagnose as a library for checkups to consume.

As an initial step, the `environment` package in defined to serve
all basic environment functionalities all checkups need.

- Read end extract the env variables.
- Extract the information about the ConfigMap.
- Read and extract the namespace name.

Use the new package in the vm-latency checkup.
As part of this change, handle the go-code dependencies between the
checkup and the kiagnose packages.

Signed-off-by: Edward Haas <edwardh@redhat.com>