GitHub actions: Run checks on release branches
Currently, our GitHub actions checks workflows (unit tests and e2e),
are triggered only on `push` or `pull_request` to the main branch.
They are not triggered on stable release branches.

Add a trigger to check workflows on `push` or `pull_request`
to release branches, in order to support back-porting as well.

Signed-off-by: Orel Misan <omisan@redhat.com>