checkups, vm-latency: Embed kiagnose into the checkup
The execution of a checkup, included two steps:
- Executing the Kiagnose Job (by the user).
- Executing the Checkup Job (by the Kiagnose application).

The vm-latency checkup is changed now to use Kiagnose as an external
library and to execute in a one step.
The logic that existed in the Kiagnose application is now imported
and used directly in the checkup.

With this change, the user is expected to execute the checkup Job
directly.
The checkup is now directly accessing the user-faced ConfigMap and the
usage of the internal ConfigMap is dropped.

Signed-off-by: Edward Haas <edwardh@redhat.com>