vmlatency: Define user at Dockerfile
Currently, the vm latency Dockerfile does not specify the user
running the kubevirt-vm-latency, thus it is implicitly running
as root.

This required the checkup Job to specify the container's user.

Add a `USER` stanza to the vm latency checkup's Dockerfile.
Remove `runAsUser` from the container's securityContext.

Install `shadow-utils` in the image in order to use the `adduser`
command, and remove it after creating the user.

Signed-off-by: Orel Misan <omisan@redhat.com>