vmlatency: Add ownerReference to source and target VMIs
Currently, if a problem occurs during the teardown process,
one or more VMIs could stay in the namespace as leftovers.

Add OwnerReference to source and target VMIs, in order to
use the automatic K8s garbage collector, so they will be
garbage-collected when the checkup Pod is deleted.

The ownerReference is only created if the Pod's name and UID
are known: The name should always be present, the UID is dependent
on the user's configuration of downward API.

Signed-off-by: Orel Misan <omisan@redhat.com>