vmlatency, client: Add context to DeleteVirtualMachineInstance()
Currently, `DeleteVirtualMachineInstance()` does not receive a context,
thus it is not cancelable from the outside.

Add context to `DeleteVirtualMachineInstance()` and create a mechanism
to artificially accept context.

This is done because KubeVirt client-go does not accept context
for `VirtualMachineInstance` methods [1].

This mechanism will be in place until the fix will be applied to
KubeVirt client-go.

[1] https://github.com/kubevirt/kubevirt/issues/8885

Signed-off-by: Orel Misan <omisan@redhat.com>