VMLatency: Refactor ParsePingResults
Currently, there is the following warning message in the
checkup's log:
```
ping parser: failed to parse 'time': time: invalid duration "ms"
```

This is caused because the checkup uses a ping
utility that does not produce the actual duration
which is expected by `ParsePingResults`.

Simplify the `ParsePingResults`:
1. The statistics pattern is unified and simplified.
2. The matching logic is simplified.

Error handling is added.

The packet loss percentage is not parsed, since it is
not part of the `Results` struct.

Signed-off-by: Orel Misan <omisan@redhat.com>